requirements.txt
numpy==2.1.2
Pillow==10.4.0
noise==1.2.2
PyYAML==6.0.2
pytest==8.3.2
black==24.8.0
ruff==0.6.8
mypy==1.11.2